[TASK] Avoid warnings of undefined index string -1 offset in Dispatcher
Depending on how the site base path is configured (if it is set or not) Dispatcher
class will try to set or append `/` to it. If the site base path will be set to an empty
string we will end up with PHP warning:

`Core: Error handler (FE): PHP Warning: Uninitialized string offset -1 in /code/vendor/aoe/restler/Classes/System/Dispatcher.php line 76` 

becasue script tries to check undefined string index.

This commit changes how the site base path is handled here. We will remove if from
URI path only if site base path:

- was not an empty string
- was not set to `/`

If it was set to an empty string or just slash, we will return URI path directly.